decouple radiance and vpnclient
