+++ radiance.go
@@ -6,16 +6,28 @@ package radiance
 import (
 	"context"
 	"fmt"
+	"io"
 	"log/slog"
+	"net/http"
+	"net/url"
+	"os"
+	"path/filepath"
 	"sync"
-	"sync/atomic"
 	"time"
 
+	"github.com/Xuanwo/go-locale"
+	"github.com/getlantern/appdir"
 	C "github.com/getlantern/common"
+	"github.com/getlantern/fronted"
+	"github.com/getlantern/kindling"
+	"github.com/sagernet/sing-box/experimental/libbox"
 
 	"github.com/getlantern/radiance/api"
+	"github.com/getlantern/radiance/app"
 	"github.com/getlantern/radiance/client"
 	"github.com/getlantern/radiance/common"
+	"github.com/getlantern/radiance/common/deviceid"
+	"github.com/getlantern/radiance/common/reporting"
 
 	boxservice "github.com/getlantern/radiance/client/service"
 	"github.com/getlantern/radiance/config"
@@ -46,28 +58,76 @@ type configHandler interface {
 
 // Radiance is a local server that proxies all requests to a remote proxy server over a transport.StreamDialer.
 type Radiance struct {
-	client.VPNClient
+	confHandler   configHandler
+	issueReporter *issue.IssueReporter
+	apiHandler    *api.APIHandler
 
-	confHandler  configHandler
-	activeServer *atomic.Value
-	stopChan     chan struct{}
 	//user config is the user config object that contains the device ID and other user data
 	userConfig common.UserInfo
+	logDir     string
+	dataDir    string
+	locale     string
 
-	issueReporter *issue.IssueReporter
-	logsDir       string
 	shutdownFuncs []func(context.Context) error
 	closeOnce     sync.Once
+	stopChan      chan struct{}
+}
+
+type Options struct {
+	DataDir  string
+	LogDir   string
+	Locale   string
+	DeviceID string
 }
 
 // NewRadiance creates a new Radiance VPN client. platIfce is the platform interface used to
 // interact with the underlying platform on iOS and Android. On other platforms, it is ignored and
 // can be nil.
-func NewRadiance(opts client.Options) (*Radiance, error) {
-	init, err := initialize(opts)
-	if err != nil {
-		return nil, fmt.Errorf("failed to initialize common: %w", err)
+func NewRadiance(opts Options) (*Radiance, error) {
+	reporting.Init()
+	if opts.DataDir == "" {
+		opts.DataDir = appdir.General(app.Name)
 	}
+	if opts.LogDir == "" {
+		opts.LogDir = appdir.Logs(app.Name)
+	}
+	if opts.Locale == "" {
+		// It is preferable to use the locale from the frontend, as locale is a requirement for lots
+		// of frontend code and therefore is more reliably supported there.
+		// However, if the frontend locale is not available, we can use the system locale as a fallback.
+		if tag, err := locale.Detect(); err != nil {
+			opts.Locale = "en-US"
+		} else {
+			opts.Locale = tag.String()
+		}
+	}
+
+	var platformDeviceID string
+	if common.IsAndroid() || common.IsIOS() {
+		platformDeviceID = opts.DeviceID
+	} else {
+		platformDeviceID = deviceid.Get()
+	}
+
+	mkdirs(opts)
+	var logWriter io.Writer
+	logWriter, err := newLog(filepath.Join(opts.LogDir, app.LogFileName))
+	if err != nil {
+		return nil, fmt.Errorf("could not create log: %w", err)
+	}
+
+	f, ferr := newFronted(logWriter, reporting.PanicListener, filepath.Join(opts.DataDir, "fronted_cache.json"))
+	if ferr != nil {
+		return nil, fmt.Errorf("failed to create fronted: %w", ferr)
+	}
+
+	k := kindling.NewKindling(
+		kindling.WithPanicListener(reporting.PanicListener),
+		kindling.WithLogWriter(logWriter),
+		kindling.WithDomainFronting(f),
+		kindling.WithProxyless("api.iantem.io"),
+	)
+
 	shutdownFuncs := []func(context.Context) error{}
 	shutdownMetrics, err := metrics.SetupOTelSDK(context.Background())
 	if err != nil {
@@ -77,50 +137,39 @@ func NewRadiance(opts client.Options) (*Radiance, error) {
 		slog.Debug("Setup OpenTelemetry SDK", "shutdown", shutdownMetrics)
 	}
 
-	vpnC, err := client.NewVPNClient(opts)
-	if err != nil {
-		slog.Error("Failed to create VPN client", "error", err)
-		return nil, fmt.Errorf("failed to create VPN client: %w", err)
-	}
-
-	u := api.NewUser(init.kindling.NewHTTPClient(), init.userConfig)
-	issueReporter, err := issue.NewIssueReporter(init.kindling.NewHTTPClient(), u, init.userConfig)
+	userConfig := common.NewUserConfig(platformDeviceID, opts.DataDir, opts.Locale)
+	u := api.NewUser(k.NewHTTPClient(), userConfig)
+	issueReporter, err := issue.NewIssueReporter(k.NewHTTPClient(), u, userConfig)
 	if err != nil {
 		return nil, fmt.Errorf("failed to create issue reporter: %w", err)
 	}
 	cOpts := config.Options{
 		PollInterval:     configPollInterval,
-		HTTPClient:       init.kindling.NewHTTPClient(),
-		User:             init.userConfig,
+		HTTPClient:       k.NewHTTPClient(),
+		User:             userConfig,
 		DataDir:          opts.DataDir,
 		ConfigRespParser: boxservice.UnmarshalConfig,
 		Locale:           opts.Locale,
 	}
 	confHandler := config.NewConfigHandler(cOpts)
-	confHandler.AddConfigListener(vpnC.OnNewConfig)
-
+	apiHandler := api.NewAPIHandlerInternal(k.NewHTTPClient(), userConfig)
 	return &Radiance{
-		VPNClient:     vpnC,
 		confHandler:   confHandler,
-		activeServer:  new(atomic.Value),
-		stopChan:      make(chan struct{}),
-		userConfig:    init.userConfig,
 		issueReporter: issueReporter,
-		logsDir:       opts.LogDir,
+		apiHandler:    apiHandler,
+		userConfig:    userConfig,
+		logDir:        opts.LogDir,
+		dataDir:       opts.DataDir,
+		locale:        opts.Locale,
 		shutdownFuncs: shutdownFuncs,
+		stopChan:      make(chan struct{}),
+		closeOnce:     sync.Once{},
 	}, nil
 }
 
-// NewAPIHandler creates a new APIHandler instance. This is used to interact with the API.
-func NewAPIHandler(opts client.Options) (*api.APIHandler, error) {
-	// Note: This separation is necessary because the iOS tunnel runs in a different isolated process.
-	// and we need access to the API in the main process.
-	init, err := initialize(opts)
-	if err != nil {
-		return nil, fmt.Errorf("failed to initialize common: %w", err)
-	}
-	apiHandler := api.NewAPIHandlerInternal(init.kindling.NewHTTPClient(), init.userConfig)
-	return apiHandler, nil
+// APIHandler returns the API handler for the Radiance client.
+func (r *Radiance) APIHandler() *api.APIHandler {
+	return r.apiHandler
 }
 
 // TODO: The server-related functionality should probably be moved to the VPNClient as well.
@@ -149,26 +198,12 @@ func (r *Radiance) Close() {
 	<-r.stopChan
 }
 
-// Server represents a remote VPN server.
-type Server struct {
-	Address  string
-	Location C.ServerLocation
-	Protocol string
-}
-
-// GetActiveServer returns the remote VPN server this client is currently connected to.
-// It returns nil when VPN is disconnected
-func (r *Radiance) GetActiveServer() (*Server, error) {
-	if !r.ConnectionStatus() {
-		return nil, fmt.Errorf("VPN is not connected")
-	}
-	activeServer := r.activeServer.Load()
-	if activeServer == nil {
-		return nil, fmt.Errorf("no active server config")
-	}
-
-	return activeServer.(*Server), nil
-
+// VPNClient returns the [client.VPNClient] instance. The client will be initialized with the provided
+// [libbox.PlatformInterface] if has not been initialized yet, or will return the existing client instance.
+// platIfce used to interact with the underlying platform on iOS and Android. On other platforms, it
+// is ignored and can be nil. enableSplitTunnel is the initial state of split tunneling when the service starts.
+func (r *Radiance) VPNClient(platIfce libbox.PlatformInterface, enableSplitTunnel bool) (client.VPNClient, error) {
+	return client.NewVPNClient(r.dataDir, r.logDir, platIfce, enableSplitTunnel)
 }
 
 // UserInfo returns the user info object for this client
@@ -228,7 +263,7 @@ func (r *Radiance) ReportIssue(email string, report *IssueReport) error {
 	}
 
 	return r.issueReporter.Report(
-		r.logsDir,
+		r.logDir,
 		email,
 		typeInt,
 		report.Description,
@@ -238,19 +273,54 @@ func (r *Radiance) ReportIssue(email string, report *IssueReport) error {
 		country)
 }
 
-func (r *Radiance) AddCustomServer(cfg boxservice.ServerConnectConfig) error {
-	return r.VPNClient.AddCustomServer(cfg)
+func mkdirs(opts Options) {
+	// Make sure the data and logs dirs exist
+	for _, dir := range []string{opts.DataDir, opts.LogDir} {
+		if err := os.MkdirAll(dir, 0o755); err != nil {
+			slog.Error("Failed to create data directory", "dir", dir, "error", err)
+		}
+	}
 }
 
-func (r *Radiance) SelectCustomServer(tag string) error {
-	return r.VPNClient.SelectCustomServer(tag)
+// Return an slog logger configured to write to both stdout and the log file.
+func newLog(logPath string) (io.Writer, error) {
+	// If the log file does not exist, create it.
+	f, err := os.OpenFile(logPath, os.O_CREATE|os.O_WRONLY|os.O_APPEND, 0o644)
+	if err != nil {
+		return nil, fmt.Errorf("failed to open log file: %w", err)
+	}
+	// defer f.Close() - file should be closed externally when logger is no longer needed
+	logWriter := io.MultiWriter(os.Stdout, f)
+	logger := slog.New(slog.NewTextHandler(logWriter, &slog.HandlerOptions{
+		AddSource: true,
+	}))
+	slog.SetDefault(logger)
+	return logWriter, nil
 }
 
-func (r *Radiance) RemoveCustomServer(tag string) error {
-	return r.VPNClient.RemoveCustomServer(tag)
-}
+func newFronted(logWriter io.Writer, panicListener func(string), cacheFile string) (fronted.Fronted, error) {
+	// Parse the domain from the URL.
+	configURL := "https://raw.githubusercontent.com/getlantern/lantern-binaries/refs/heads/main/fronted.yaml.gz"
+	u, err := url.Parse(configURL)
+	if err != nil {
+		return nil, fmt.Errorf("failed to parse URL: %v", err)
+	}
+	// Extract the domain from the URL.
+	domain := u.Host
 
-// SplitTunnelHandler returns the split tunnel handler for the VPN client.
-func (r *Radiance) SplitTunnelHandler() *client.SplitTunnel {
-	return r.VPNClient.SplitTunnelHandler()
+	// First, download the file from the specified URL using the smart dialer.
+	// Then, create a new fronted instance with the downloaded file.
+	trans, err := kindling.NewSmartHTTPTransport(logWriter, domain)
+	if err != nil {
+		return nil, fmt.Errorf("failed to create smart HTTP transport: %v", err)
+	}
+	httpClient := &http.Client{
+		Transport: trans,
+	}
+	return fronted.NewFronted(
+		fronted.WithPanicListener(panicListener),
+		fronted.WithCacheFile(cacheFile),
+		fronted.WithHTTPClient(httpClient),
+		fronted.WithConfigURL(configURL),
+	), nil
 }
